Fixed the chk file being deleted before it could be used in an auto report to generate NTOs

+++ silico/submit/program/gaussian.py
@@ -143,6 +143,7 @@ class Gaussian(Program_target):
             Post submission method.
             """
             # Chk/fchk management. Do this before making report (in super()) to avoid making fchk twice.
+            chk_conversion_success = True
             try:
                 # Create an fchk file if asked.
                 if self.calculation.convert_chk:
@@ -150,9 +151,15 @@ class Gaussian(Program_target):
                     fchk_file.get_file()
             except Exception:
                 silico.logging.get_logger().error("Failed to create fchk file", exc_info = True)
-            else:
+                chk_conversion_success = False
+            
+            # Use our parent to create result and report files.
+            super().post()
+                
+            # Now process results and reports (in case we need .chk or .rwf file, which will be deleted in but a moment).
+            if chk_conversion_success:
                 try:
-                    # Now delete the chk file if we were asked to.
+                    # Now delete the chk file if we were asked to (but only if the conversion to fchk was successful).
                     if not self.calculation.keep_chk:
                         self.chk_file_path.unlink()
                 except FileNotFoundError:
@@ -161,8 +168,9 @@ class Gaussian(Program_target):
                 except Exception:
                     silico.logging.get_logger().error("Failed to delete chk file", exc_info = True)
             
-            # Use our parent to create result and report files.
-            super().post()
+            else:
+                # Let the user know we're ignoring their request.
+                silico.logging.get_logger().warning("Not deleting chk file because conversion to fchk was not successful")
             
             # Remove rwf if we've been asked to.
             try:
